+++ internal/cli/http/client.go
@@ -7,27 +7,46 @@ import (
 	"os"
 )
 
-var (
-	Client     http.Client
-	ApiLogin   string
-	ApiSignUp  string
-	ApiAddItem string
-	ApiGetItem string
+const (
+	apiLogin   = "/user/login"
+	apiSignUp  = "/user/signup"
+	apiAddItem = "/item/add"
+	apiGetItem = "/item"
+	apiBase    = "/api/v1"
 )
 
-func InitClient(certPath string, baseURL string) error {
-	apiURL := baseURL + "/api/v1"
-	ApiLogin = apiURL + "/user/login"
-	ApiSignUp = apiURL + "/user/signup"
-	ApiAddItem = apiURL + "/item/add"
-	ApiGetItem = apiURL + "/item"
+type Api struct {
+	Login   string
+	SignUp  string
+	AddItem string
+	GetItem string
+}
+
+type Client struct {
+	token   string
+	baseURL string
+	Api     Api
+	http.Client
+}
+
+func InitClient(certPath string, baseURL string) (*Client, error) {
+	apiURL := baseURL + apiBase
+	api := Api{
+		Login:   apiURL + apiLogin,
+		SignUp:  apiURL + apiSignUp,
+		AddItem: apiURL + apiAddItem,
+		GetItem: apiURL + apiGetItem,
+	}
 	if certPath == "" {
-		Client = http.Client{}
-		return nil
+		return &Client{
+			Client:  http.Client{},
+			baseURL: baseURL,
+			Api:     api,
+		}, nil
 	}
 	caCert, err := os.ReadFile(certPath)
 	if err != nil {
-		return err
+		return nil, err
 	}
 	caCertPool := x509.NewCertPool()
 	caCertPool.AppendCertsFromPEM(caCert)
@@ -35,13 +54,15 @@ func InitClient(certPath string, baseURL string) error {
 	// т.к. Wireshark показывает TLS и данные выглядят зашифрованными.
 	// Ошибка "failed to verify certificate" возникает, если создать пустой CertPool и не грузить туда ничего.
 	// Если создать дефолтный http.Client{} без CertPool, то по дефолту какой-то сертификат видимо есть.
-
-	Client = http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				RootCAs: caCertPool,
+	return &Client{
+		Client: http.Client{
+			Transport: &http.Transport{
+				TLSClientConfig: &tls.Config{
+					RootCAs: caCertPool,
+				},
 			},
 		},
-	}
-	return nil
+		baseURL: baseURL,
+		Api:     api,
+	}, nil
 }
